Move metrics to db package: it makes more sense
