+++ pkg/db/metrics.go
@@ -1,4 +1,4 @@
-package analyzer
+package db
 
 import (
 	"fmt"
@@ -14,6 +14,10 @@ type DBMetrics struct {
 	ValidatorRewards    bool
 	Withdrawals         bool
 	Transactions        bool
+
+	BlockDownload   bool
+	StateDownload   bool
+	RewardsDownload bool
 }
 
 func NewMetrics(input string) (DBMetrics, error) {
@@ -24,18 +28,25 @@ func NewMetrics(input string) (DBMetrics, error) {
 		switch item {
 		case "block":
 			dbMetrics.Block = true
+			dbMetrics.BlockDownload = true
 		case "epoch":
 			dbMetrics.Epoch = true
+			dbMetrics.StateDownload = true
 		case "pool_summary":
 			dbMetrics.PoolSummary = true
 		case "validator_last_status":
 			dbMetrics.ValidatorLastStatus = true
+			dbMetrics.StateDownload = true
 		case "validator":
 			dbMetrics.ValidatorRewards = true
+			dbMetrics.StateDownload = true
+			dbMetrics.RewardsDownload = true
 		case "withdrawals":
 			dbMetrics.Withdrawals = true
+			dbMetrics.BlockDownload = true
 		case "transactions":
 			dbMetrics.Transactions = true
+			dbMetrics.BlockDownload = true
 		default:
 			return DBMetrics{}, fmt.Errorf("could not parse metric: %s", item)
 		}
